rd: make SparseRDGraphBuilder abstract
move implementation to CytronSRGBuilder
